backend/requirements.txt
alembic==1.13.2
asgiref==3.8.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==4.1.3
blinker==1.9.0
certifi==2020.6.20
cffi==1.17.1
chardet==3.0.4
click==8.1.7
cloud-init==25.1.4
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==43.0.3
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1.2
django-cors-headers==4.5.0
django-paypal==2.1
djangorestframework==3.15.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.8
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.0.3
git-filter-repo==2.45.0
greenlet==3.0.3
gunicorn==22.0.0
gyp==0.1
httplib2==0.20.2
hyperlink==21.0.0
idna==2.6
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Mako==1.3.5
MarkupSafe==3.0.2
more-itertools==8.10.0
mysql-connector-python==9.4.0
mysqlclient==2.1.1
netifaces==0.11.0
numpy==2.1.0
oauthlib==3.2.0
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
paypalrestsdk==1.13.3
psycopg2-binary==2.9.9
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycodestyle==2.6.0
pycparser==2.22
pycurl==7.44.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==24.2.1
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==5.4.1
requests==2.18.4
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
SQLAlchemy==2.0.31
sqlparse==0.4.2
stripe==11.1.1
systemd-python==234
texttable==1.6.4
Twisted==22.1.0
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.22
wadllib==1.3.6
websocket-client==1.2.3
Werkzeug==3.1.3
zipp==1.0.0
zope.interface==5.4.0
